+++ Package.swift
@@ -1,4 +1,4 @@
-// swift-tools-version:5.0
+// swift-tools-version:5.3
 
 import PackageDescription
 
@@ -8,7 +8,7 @@ let package = Package(
         .library(name: "SocketIO", targets: ["SocketIO"])
     ],
     dependencies: [
-        .package(url: "https://github.com/daltoniam/Starscream", .upToNextMinor(from: "3.1.0")),
+        .package(url: "https://github.com/daltoniam/Starscream", .upToNextMinor(from: "4.0.0")),
     ],
     targets: [
         .target(name: "SocketIO", dependencies: ["Starscream"]),