+++ Source/SocketIO/Manager/SocketManagerSpec.swift
@@ -45,7 +45,6 @@ import Foundation
 /// To disconnect a socket and remove it from the manager, either call `SocketIOClient.disconnect()` on the socket,
 /// or call one of the `disconnectSocket` methods on this class.
 ///
-@objc
 public protocol SocketManagerSpec : AnyObject, SocketEngineClient {
     // MARK: Properties
 
@@ -71,7 +70,7 @@ public protocol SocketManagerSpec : AnyObject, SocketEngineClient {
 
     /// The minimum number of seconds to wait before attempting to reconnect.
     var reconnectWait: Int { get set }
-    
+
     /// The maximum number of seconds to wait before attempting to reconnect.
     var reconnectWaitMax: Int { get set }
 
@@ -84,6 +83,9 @@ public protocol SocketManagerSpec : AnyObject, SocketEngineClient {
     /// The status of this manager.
     var status: SocketIOStatus { get }
 
+    /// The version of socket.io in use.
+    var version: SocketIOVersion { get }
+
     // MARK: Methods
 
     /// Connects the underlying transport.
@@ -92,7 +94,8 @@ public protocol SocketManagerSpec : AnyObject, SocketEngineClient {
     /// Connects a socket through this manager's engine.
     ///
     /// - parameter socket: The socket who we should connect through this manager.
-    func connectSocket(_ socket: SocketIOClient)
+    /// - parameter withPayload: Optional payload to send on connect
+    func connectSocket(_ socket: SocketIOClient, withPayload: [String: Any]?)
 
     /// Called when the manager has disconnected from socket.io.
     ///
@@ -116,7 +119,7 @@ public protocol SocketManagerSpec : AnyObject, SocketEngineClient {
     ///
     /// - parameter event: The event to send.
     /// - parameter items: The data to send with this event.
-    func emitAll(_ event: String, withItems items: [Any])
+    func emitAll(_ event: String, _ items: SocketData...)
 
     /// Tries to reconnect to the server.
     ///